+++ src/APIAggregator.API/Features/ExternalAPIs/ILocationDataProvider.cs
@@ -3,6 +3,6 @@
 	public interface ILocationDataProvider
 	{
 		string Name { get; }
-		Task<object> GetDataAsync(float latitude, float longitude, CancellationToken cancellationToken);
+		Task<object> GetDataAsync(double latitude, double longitude, CancellationToken cancellationToken);
 	}
 }
